fix(dragdrop): Apply styling code review
