+++ src/Uno.UWP/ApplicationModel/DataTransfer/DragDrop/Core/IDragEventSource.cs
@@ -1,0 +1,19 @@
+#nullable enable
+
+using System;
+using System.Linq;
+using Windows.Foundation;
+
+namespace Windows.ApplicationModel.DataTransfer.DragDrop.Core
+{
+	internal interface IDragEventSource
+	{
+		long Id { get; }
+
+		uint FrameId { get; }
+
+		(Point location, DragDropModifiers modifier) GetState();
+
+		Point GetPosition(object? relativeTo);
+	}
+}
